+++ cpan/attributes/default.rb
@@ -1,4 +1,12 @@
+default.cpan_client.bootstrap.packages = ['curl']
+
+case platforms
+when 'centos'
+    default.cpan_client.bootstrap.packages << 'perl-devel'
+end
+
+default.cpan_client.bootstrap.cpan_packages = ['Time::HiRes', 'CPAN', 'local::lib', 'App::pmuninstall']
+
 default.cpan_client.default_inc = []
-default.cpan_client.bootstrap.install_base = nil
-default.cpan_client.bootstrap.deps = ['CPAN', 'local::lib', 'App::pmuninstall']
+