+++ cpan/metadata.rb
@@ -4,8 +4,8 @@ maintainer_email "melezhik@gmail.com"
 license          "All rights reserved"
 description      "CPAN modules provider (cpan_client provider LWRP)"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.0.26"
-%w{ ubuntu gentoo debian }.each do |os|
+version          "0.0.27"
+%w{ debian ubuntu centos  gentoo  }.each do |os|
   supports os
 end