+++ cpan/recipes/bootstrap.rb
@@ -1,7 +1,10 @@
 include_recipe 'cpan'
-package 'curl'
+
+node.cpan_client.bootstrap.packages.each  { |p| package p } 
+
 execute 'curl -L http://cpanmin.us | perl - --sudo App::cpanminus'
-node.cpan_client.bootstrap.deps.each  do |m|
+
+node.cpan_client.bootstrap.cpan_packages.each  do |m|
     execute "cpanm #{m}" do
         user 'root'
         group 'root'